Ignore lines with leading / after trailing strings
Two configurations were added:

g:yadi_ignore_after_trailing
If the preceding line of current one ends with any of the strings in
this list, current line is ignored in DetectIndent.

g:yadi_ignore_with_leading
If current line begins with any of the strings in this list, current
line is ignored in DetectIndent.

Fixes #1.